Rating bars are now set to show up to 3 stars max

+++ app/src/main/java/fr/azhot/go4lunch/view/RestaurantDetailActivity.java
@@ -29,7 +29,7 @@ public class RestaurantDetailActivity extends AppCompatActivity {
         String name = intent.getStringExtra("name");
         String details = intent.getStringExtra("details");
         String photoUrl = intent.getStringExtra("photoUrl");
-        float rating = intent.getFloatExtra("rating", 0);
+        int rating = intent.getIntExtra("rating", 0);
 
         mBinding.restaurantNameTextView.setText(name);
         mBinding.restaurantDetails.setText(details);
